Partial rewrite of game in C w/ ncurses
To expand the game from the web to the command line, I've started a
rewriting effort for the game using ncurses. The logic of the game
should be nearly identical, but will probably forego audio for this
version.

The refactoring effort is only partially complete at this point, and
still needs UI updates such as coloring, and a wider variety of
obstacles. That being said, the game is playable in this state.