Reset game on player death
This modifies the game loop logic to reset the game when the player
dies, rather than aborting the game completely. The "main menu" is
updated with a message indicating player death, to reduce any sort of
ambiguity for why the game has been reset.

Also introduces the game's first (and likely only) "goto" call, for
breaking out of the nested game loop cleanly on death.